requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
babel==2.16.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
django==4.2.11
django-allauth==0.63.3
django-celery-beat
django-cors-headers==4.6.0
django-countries==7.6.1
django-extensions==3.2.3
django-filter==24.2
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
factory-boy==3.3.0
faker==33.3.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
pillow
promise==2.3
prompt-toolkit==3.0.48
psutil==6.1.1
psycopg2-binary
py-money==0.5.0
pyjwt==2.8.0
pymemcache==4.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
text-unidecode==1.3
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
